scoring_api/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
scikit-learn==1.3.2
pandas==2.1.4
joblib==1.3.2
google-cloud-storage==2.18.2

